pyproject.toml
[project]
name = "ocr-vector-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langchain-openai>=0.0.5",
    "langchain-pinecone>=0.1.0",
    "tiktoken>=0.5.0",
    "openai>=1.12.0",
    "pinecone-client>=3.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf4llm>=0.0.27",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
]
